+++ Resources/Prototypes/_Trauma/StatusIcon/job.yml
@@ -2,7 +2,7 @@
   parent: JobIcon
   id: JobIconPhoneMerchant
   icon:
-    sprite: /Textures/_Goobstation/Interface/Misc/job_icons.rsi
+    sprite: /Textures/_Trauma/Interface/Misc/job_icons.rsi
     state: phonemerchant
   jobName: job-name-phonemerchant
   allowSelection: false
