+++ Resources/Prototypes/_Wega/Entities/Objects/Devices/pda.yml
@@ -28,5 +28,5 @@
   - type: Pda
     id: PhoneMerchantIDCard
   - type: PdaBorderColor
-    borderColor: "#000000"
+    borderColor: "#c48c4d"
     accentVColor: "#C0C0C0"
