+++ Resources/Textures/Objects/Misc/id_cards.rsi/meta.json
@@ -1,7 +1,7 @@
 {
   "version": 1,
   "license": "CC-BY-SA-3.0",
-  "copyright": "Taken from tgstation at commit https://github.com/tgstation/tgstation/commit/d917f4c2a088419d5c3aec7656b7ff8cebd1822e idcluwne made by brainfood1183 (github) for ss14, idbrigmedic made by PuroSlavKing (Github), pirate made by brainfood1183 (github), idadmin made by Arimah (github), idvisitor by IProduceWidgets (Github), idintern-service by spanky-spanky (Github), Chemist made by Lehmin for Goobstation, idpartymaker idsecurityclown idcommandmaid by lzk228 (github), idradiohost and phonemerchant made by john john (github). secclown Recolored by Struater",
+  "copyright": "Taken from tgstation at commit https://github.com/tgstation/tgstation/commit/d917f4c2a088419d5c3aec7656b7ff8cebd1822e idcluwne made by brainfood1183 (github) for ss14, idbrigmedic made by PuroSlavKing (Github), pirate made by brainfood1183 (github), idadmin made by Arimah (github), idvisitor by IProduceWidgets (Github), idintern-service by spanky-spanky (Github), Chemist made by Lehmin for Goobstation, idpartymaker idsecurityclown idcommandmaid by lzk228 (github), idradiohost made by john john (github). secclown Recolored by Struater",
   "size": {
     "x": 32,
     "y": 32
@@ -168,9 +168,6 @@
     },
     {
       "name": "idradiohost"
-    },
-    {
-      "name": "idphonemerchant"
     },
     {
       "name": "idzookeeper"
