+++ Resources/Textures/_Goobstation/Interface/Misc/job_icons.rsi/meta.json
@@ -1,7 +1,7 @@
 {
     "version": 1,
     "license": "CC-BY-SA-3.0",
-    "copyright": "Future taken from space stories  https://github.com/MetalSage/space-stories-archive/commit/26812e79c19c5f998bcbc445afbdceea625bc887 Blueshield, Nanotrasen Representative, and Trainer taken from Paradise @ https://github.com/ParadiseSS13/Paradise/commit/5a68c5f6d3b60ee82c06e0287d1eb8108d2e1fe2, Lambda & Xenobiologist by SolsticeOfTheWinter, radio host and phone merchant made by John John",
+    "copyright": "Future taken from space stories  https://github.com/MetalSage/space-stories-archive/commit/26812e79c19c5f998bcbc445afbdceea625bc887 Blueshield, Nanotrasen Representative, and Trainer taken from Paradise @ https://github.com/ParadiseSS13/Paradise/commit/5a68c5f6d3b60ee82c06e0287d1eb8108d2e1fe2, Lambda & Xenobiologist by SolsticeOfTheWinter, radio host made by John John",
     "size": {
         "x": 8,
         "y": 8
@@ -41,9 +41,6 @@
     },
     {
         "name": "radiohost"
-    },
-    {
-        "name": "phonemerchant"
     }
     ]
 }
