+++ Resources/Textures/_Trauma/Interface/Misc/job_icons.rsi/meta.json
@@ -1,0 +1,17 @@
+{
+    "version": 1,
+    "license": "CC-BY-SA-3.0",
+    "copyright": "Taken from https://github.com/tgstation/tgstation/blob/ce6beb8a4d61235d9a597a7126c407160ed674ea/icons/mob/huds/hud.dmi and recoloured by @deltanedas (github), phonemerchant made by John John",
+    "size": {
+        "x": 8,
+        "y": 8
+    },
+    "states": [
+        {
+            "name": "Geneticist"
+        },
+        {
+            "name": "Phonemerchant"
+        }
+    ]
+}
